# tests/contract/snapshots/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "ModelCatalogResponse": {
        "additionalProperties": false,
        "properties": {
          "models": {
            "items": {
              "type": "string"
            },
            "title": "Models",
            "type": "array"
          }
        },
        "required": [
          "models"
        ],
        "title": "ModelCatalogResponse",
        "type": "object"
      },
...
      "ModelProfileCreate": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/ModelKind"
          },
          "model": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Model",
            "type": "string"
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "parameters": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChatModelParameters"
              },
              {
                "$ref": "#/components/schemas/EmbeddingModelParameters"
              }
            ],
            "title": "Parameters"
          },
          "provider_id": {
            "format": "uuid",
            "title": "Provider Id",
            "type": "string"
          }
        },
        "required": [
          "provider_id",
          "name",
          "kind",
          "model",
          "parameters"
        ],
        "title": "ModelProfileCreate",
        "type": "object"
      },
      "ModelProfileResponse": {
        "additionalProperties": false,
        "properties": {
          "capability_fingerprint": {
            "title": "Capability Fingerprint",
            "type": "string"
          },
          "compatibility_fingerprint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compatibility Fingerprint"
          },
          "configuration_fingerprint": {
            "title": "Configuration Fingerprint",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "embedding_validation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EmbeddingValidationResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ModelKind"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "parameters": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChatModelParameters"
              },
              {
                "$ref": "#/components/schemas/EmbeddingModelParameters"
              }
            ],
            "title": "Parameters"
          },
          "provider_id": {
            "format": "uuid",
            "title": "Provider Id",
            "type": "string"
          },
          "provider_revision_id": {
            "format": "uuid",
            "title": "Provider Revision Id",
            "type": "string"
          },
          "provider_secret_available": {
            "title": "Provider Secret Available",
            "type": "boolean"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "revision_id": {
            "format": "uuid",
            "title": "Revision Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "validated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validated At"
          },
          "validation_error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Error Code"
          },
          "validation_status": {
            "$ref": "#/components/schemas/ModelValidationStatus"
          }
        },
        "required": [
          "id",
          "revision_id",
          "revision",
          "provider_id",
          "provider_revision_id",
          "name",
          "kind",
          "model",
          "parameters",
          "enabled",
          "provider_secret_available",
          "validation_status",
          "validation_error_code",
          "validated_at",
          "configuration_fingerprint",
          "capability_fingerprint",
          "compatibility_fingerprint",
          "created_at",
          "updated_at"
        ],
        "title": "ModelProfileResponse",
        "type": "object"
      },
      "ModelProfileUpdate": {
        "additionalProperties": false,
        "properties": {
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "model": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "parameters": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChatModelParameters"
              },
              {
                "$ref": "#/components/schemas/EmbeddingModelParameters"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parameters"
          },
          "provider_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Id"
          }
        },
        "title": "ModelProfileUpdate",
        "type": "object"
      },
      "ModelProviderCreate": {
        "additionalProperties": false,
        "properties": {
          "api_key": {
            "format": "password",
            "title": "Api Key",
            "type": "string",
            "writeOnly": true
          },
          "base_url": {
            "format": "uri",
            "minLength": 1,
            "title": "Base Url",
            "type": "string"
          },
          "max_concurrency": {
            "default": 2,
            "maximum": 32.0,
            "minimum": 1.0,
            "title": "Max Concurrency",
            "type": "integer"
          },
          "max_retries": {
            "default": 1,
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Max Retries",
            "type": "integer"
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "protocol": {
            "$ref": "#/components/schemas/ModelProviderProtocol"
          },
          "timeout_seconds": {
            "default": 60.0,
            "exclusiveMinimum": 0.0,
            "maximum": 600.0,
            "title": "Timeout Seconds",
            "type": "number"
          }
        },
        "required": [
          "name",
          "protocol",
          "base_url",
          "api_key"
        ],
        "title": "ModelProviderCreate",
        "type": "object"
      },
...
      "ModelSelectionUpdate": {
        "additionalProperties": false,
        "properties": {
          "chat_profile_revision_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chat Profile Revision Id"
          },
          "multimodal_embedding_profile_revision_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Multimodal Embedding Profile Revision Id"
          },
          "text_embedding_profile_revision_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Embedding Profile Revision Id"
          }
        },
        "title": "ModelSelectionUpdate",
        "type": "object"
      },
...
      "ProblemDetails": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldViolation"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "instance": {
            "title": "Instance",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "status": {
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "trace_id",
          "retryable"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Enterprise Knowledge Base API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/model-profiles": {
      "post": {
        "operationId": "create_model_profile_api_v1_model_profiles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelProfileCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Idempotency or resource conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Create Model Profile",
        "tags": [
          "model-settings"
        ]
      }
    },
    "/api/v1/model-profiles/{profile_id}": {
      "patch": {
        "operationId": "update_model_profile_api_v1_model_profiles__profile_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Idempotency or resource conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Update Model Profile",
        "tags": [
          "model-settings"
        ]
      }
    },
    "/api/v1/model-profiles/{profile_id}/validate": {
      "post": {
        "operationId": "validate_model_profile_api_v1_model_profiles__profile_id__validate_post",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Idempotency or resource conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Validate Model Profile",
        "tags": [
          "model-settings"
        ]
      }
    },
    "/api/v1/model-providers": {
      "post": {
        "operationId": "create_model_provider_api_v1_model_providers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelProviderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelProviderResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Idempotency or resource conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Create Model Provider",
        "tags": [
          "model-settings"
        ]
      }
    },
    "/api/v1/model-providers/{provider_id}": {
      "patch": {
        "operationId": "update_model_provider_api_v1_model_providers__provider_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Provider Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelProviderUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelProviderResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Idempotency or resource conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Update Model Provider",
        "tags": [
          "model-settings"
        ]
      }
    },
    "/api/v1/model-providers/{provider_id}/models": {
      "get": {
        "operationId": "list_provider_models_api_v1_model_providers__provider_id__models_get",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Provider Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelCatalogResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Idempotency or resource conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "List Provider Models",
        "tags": [
          "model-settings"
        ]
      }
    },
    "/api/v1/model-selection": {
      "put": {
        "operationId": "update_model_selection_api_v1_model_selection_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelSelectionUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelSelectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Idempotency or resource conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Update Model Selection",
        "tags": [
          "model-settings"
        ]
      }
    },
    "/api/v1/model-settings": {
      "get": {
        "operationId": "get_model_settings_api_v1_model_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Get Model Settings",
        "tags": [
          "model-settings"
        ]
      }
    },
...
  }
}